+++ Game/Player.h
@@ -16,7 +16,7 @@ private :
 	void move(Direction direction, float elapsed);
 
 public:
-	Player(float x, float y, const char* imagePath);
+	Player(int x, int y, const char* imagePath);
 	~Player();
 
 	void update(float elapsed);
